rename + add USDC updater
