+++ packages/bots/pyth-updater/src/config/zkevm.ts
@@ -8,4 +8,11 @@ export const pythConfig: PythAssetConfig[] = [
     validTimePeriodSeconds: 86400, // 24 hours
     deviationThresholdBps: 200, // 1%
   },
+  {
+    // USDC/USD
+    priceId: '0xeaa020c61cc479712813461ce153894a96a6c00b21ed0cfc2798d1f9a9e9c94a',
+    configRefreshRateInSeconds: 3600,
+    validTimePeriodSeconds: 86400, // 24 hours
+    deviationThresholdBps: 100, // 1%
+  },
 ];
